code simplified
