+++ lib/ExternalModule.js
@@ -294,20 +294,19 @@ const getSourceForScriptExternal = (urlAndGlobal, runtimeTemplate) => {
 			expression: `new Promise(${runtimeTemplate.basicFunction(
 				"resolve, reject",
 				[
-					`if(!${RuntimeGlobals.shareScopeMap} ||
-						!${RuntimeGlobals.hasOwnProperty}(${RuntimeGlobals.shareScopeMap}, 'default') || 
-						!${RuntimeGlobals.hasOwnProperty}(${RuntimeGlobals.shareScopeMap}['default'], 'webpack/container/reference/${globalName}')) {
+					`if(!${RuntimeGlobals.shareScopeMap} || !${RuntimeGlobals.hasOwnProperty}(${RuntimeGlobals.shareScopeMap}, 'default')) {
 						console.error('Host reference ${globalName} was not loaded');
-						reject(__webpack_error__);
-					}
+						reject(__webpack_error__);						
+					};
 					resolve();`
 				]
 			)}).then(${runtimeTemplate.returningFunction(
-				`
-				{get: ${runtimeTemplate.basicFunction("module, getScope", [
-					`var [[scope]] = getScope; return ${RuntimeGlobals.shareScopeMap}[scope][module][0].get();`
-				])}}
-				`
+				`{get: ${runtimeTemplate.basicFunction("module, getScope", [
+					`var [[scope]] = getScope;
+					const m = '${globalName}' + module.replace('.', '');
+					if(!${RuntimeGlobals.shareScopeMap}[scope][m]) throw new Error("Module not found in the host");			
+					return ${RuntimeGlobals.shareScopeMap}[scope][m][0].get();`
+				])}}`
 			)})`,
 			runtimeRequirements: RUNTIME_REQUIREMENTS_FOR_SCRIPT
 		};
