+++ lib/container/ContainerEntryDependency.js
@@ -15,14 +15,12 @@ class ContainerEntryDependency extends Dependency {
 	 * @param {string} name entry name
 	 * @param {[string, ExposeOptions][]} exposes list of exposed modules
 	 * @param {string} shareScope name of the share scope
-	 * @param {[string][]} remotes list of remotes
 	 */
-	constructor(name, exposes, shareScope, remotes) {
+	constructor(name, exposes, shareScope) {
 		super();
 		this.name = name;
 		this.exposes = exposes;
 		this.shareScope = shareScope;
-		this.remotes = remotes;
 	}
 
 	/**
