+++ lib/container/ContainerEntryModule.js
@@ -42,14 +42,12 @@ class ContainerEntryModule extends Module {
 	 * @param {string} name container entry name
 	 * @param {[string, ExposeOptions][]} exposes list of exposed modules
 	 * @param {string} shareScope name of the share scope
-	 * @param {[string][]} remotes list of remotes
 	 */
-	constructor(name, exposes, shareScope, remotes) {
+	constructor(name, exposes, shareScope) {
 		super("javascript/dynamic", null);
 		this._name = name;
 		this._exposes = exposes;
 		this._shareScope = shareScope;
-		this._remotes = remotes;
 	}
 
 	/**
@@ -200,19 +198,6 @@ class ContainerEntryModule extends Module {
 			);
 		}
 
-		const exposesHost = [];
-		if (this._remotes) {
-			this._remotes.forEach(remoteKey => {
-				exposesHost.push(`
-				Object.keys(shareScope).forEach((shareScopeKey) => {
-					if(!shareScopeKey.includes('webpack/container/reference/${remoteKey}')) return;
-					${RuntimeGlobals.moduleFactories}[shareScopeKey.replace('reference', 'remote')] = module => {
-						module.exports = shareScope[shareScopeKey][0].get()();
-					};
-				});`);
-			});
-		}
-
 		const source = Template.asString([
 			`var moduleMap = {`,
 			Template.indent(getters.join(",\n")),
@@ -240,8 +225,8 @@ class ContainerEntryModule extends Module {
 				`var name = ${JSON.stringify(this._shareScope)}`,
 				`var oldScope = ${RuntimeGlobals.shareScopeMap}[name];`,
 				"// when its already loaded by entrypoint, this will be already initialized, and shared references are global",
-				`if(oldScope) return oldScope`,
-				exposesHost.join("\n"),
+				// `if(oldScope) return oldScope`,
+				`if(oldScope && oldScope !== shareScope) throw new Error("Container initialization failed as it has already been initialized with a different share scope");`,
 				`${RuntimeGlobals.shareScopeMap}[name] = shareScope;`,
 				`return ${RuntimeGlobals.initializeSharing}(name, initScope);`
 			])};`,
@@ -286,7 +271,7 @@ class ContainerEntryModule extends Module {
 
 	static deserialize(context) {
 		const { read } = context;
-		const obj = new ContainerEntryModule(read(), read(), read(), read());
+		const obj = new ContainerEntryModule(read(), read(), read());
 		obj.deserialize(context);
 		return obj;
 	}
