+++ lib/container/ContainerEntryModuleFactory.js
@@ -21,12 +21,7 @@ module.exports = class ContainerEntryModuleFactory extends ModuleFactory {
 	create({ dependencies: [dependency] }, callback) {
 		const dep = /** @type {ContainerEntryDependency} */ (dependency);
 		callback(null, {
-			module: new ContainerEntryModule(
-				dep.name,
-				dep.exposes,
-				dep.shareScope,
-				dep.remotes
-			)
+			module: new ContainerEntryModule(dep.name, dep.exposes, dep.shareScope)
 		});
 	}
 };
