+++ lib/container/ContainerPlugin.js
@@ -5,6 +5,7 @@
 
 "use strict";
 
+const SharePlugin = require("../sharing/SharePlugin");
 const createSchemaValidation = require("../util/create-schema-validation");
 const ContainerEntryDependency = require("./ContainerEntryDependency");
 const ContainerEntryModuleFactory = require("./ContainerEntryModuleFactory");
@@ -32,6 +33,18 @@ class ContainerPlugin {
 	constructor(options) {
 		validate(options);
 
+		const sharedExposes = Object.keys(options.exposes).reduce(
+			(items, exposeKey) => ({
+				...items,
+				[options.exposes[exposeKey]]: {
+					singleton: true,
+					eager: true,
+					shareKey: `${options.name}/${exposeKey.replace(/^(.\/|\/)/gim, "")}`
+				}
+			}),
+			{}
+		);
+
 		this._options = {
 			name: options.name,
 			shareScope: options.shareScope || "default",
@@ -52,7 +65,7 @@ class ContainerPlugin {
 					name: item.name || undefined
 				})
 			),
-			remotes: options.remotes
+			sharedExposes
 		};
 	}
 
@@ -62,18 +75,27 @@ class ContainerPlugin {
 	 * @returns {void}
 	 */
 	apply(compiler) {
-		const { name, exposes, shareScope, filename, library, runtime, remotes } =
-			this._options;
+		const {
+			name,
+			exposes,
+			shareScope,
+			filename,
+			library,
+			runtime,
+			sharedExposes
+		} = this._options;
+
+		if (Object.keys(sharedExposes).length > 0) {
+			new SharePlugin({
+				shared: sharedExposes,
+				shareScope
+			}).apply(compiler);
+		}
 
 		compiler.options.output.enabledLibraryTypes.push(library.type);
 
 		compiler.hooks.make.tapAsync(PLUGIN_NAME, (compilation, callback) => {
-			const dep = new ContainerEntryDependency(
-				name,
-				exposes,
-				shareScope,
-				remotes
-			);
+			const dep = new ContainerEntryDependency(name, exposes, shareScope);
 			dep.loc = { name };
 			compilation.addEntry(
 				compilation.options.context,
