+++ lib/container/ModuleFederationPlugin.js
@@ -65,25 +65,7 @@ class ModuleFederationPlugin {
 					library,
 					filename: options.filename,
 					runtime: options.runtime,
-					exposes: options.exposes,
-					remotes: Object.keys(options.remotes || {})
-				}).apply(compiler);
-				const sharedExposes = Object.keys(options.exposes).reduce(
-					(items, exposeKey) => ({
-						...items,
-						[options.exposes[exposeKey]]: {
-							singleton: true,
-							eager: true,
-							shareKey: `webpack/container/reference/${
-								options.name
-							}/${exposeKey.replace(/^(.\/|\/)/gim, "")}`
-						}
-					}),
-					{}
-				);
-				new SharePlugin({
-					shared: sharedExposes,
-					shareScope: options.shareScope
+					exposes: options.exposes
 				}).apply(compiler);
 			}
 			if (
