+++ schemas/plugins/container/ContainerPlugin.json
@@ -273,10 +273,6 @@
       "type": "string",
       "minLength": 1
     },
-    "remotes": {
-      "description": "List of remotes.",
-      "type": "array"
-    },
     "runtime": {
       "$ref": "#/definitions/EntryRuntime"
     },
