+++ types.d.ts
@@ -2413,11 +2413,6 @@ declare interface ContainerPluginOptions {
 	name: string;
 
 	/**
-	 * List of remotes.
-	 */
-	remotes?: any[];
-
-	/**
 	 * The name of the runtime chunk. If set a runtime chunk with this name is created or an existing entrypoint is used as runtime.
 	 */
 	runtime?: string | false;
